requirements.txt
pdfplumber
pymupdf
moviepy
pyttsx3
openai
requests
python-dotenv
piper-tts
